extract out tables from incident.py
