+++ backend/database/__init__.py
@@ -20,5 +20,8 @@ from .models.multimedia import *
 from .models.officer import *
 from .models.participant import *
 from .models.tag import *
+from .models.result_of_stop import *
+from .models.action import *
+from .models.use_of_force import *
 # from .models.users import *
 from .models.victim import *
