+++ backend/database/models/use_of_force.py
@@ -1,0 +1,7 @@
+from .. import db
+
+
+class UseOfForce(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    incident_id = db.Column(db.Integer, db.ForeignKey("incident.id"))
+    item = db.Text()
